+++ packages/dgt-id-proxy/lib/public-api.ts
@@ -1,6 +1,7 @@
 export * from './server/auth-state-request.handler';
 export * from './server/auth-state-response.handler';
 export * from './server/client-compression.handler';
+export * from './server/client-credentials-path.handler';
 export * from './server/client-id-dynamic-auth-request.handler';
 export * from './server/client-id-dynamic-token.handler';
 export * from './server/client-id-static-auth-request.handler';
@@ -17,12 +18,17 @@ export * from './server/pkce-auth-request.handler';
 export * from './server/pkce-code-response.handler';
 export * from './server/pkce-token.handler';
 export * from './server/solid-audience-response.handler';
+export * from './server/auth0-client-redirect.handler';
+export * from './server/auth0-login-state.handler';
+export * from './server/auth0-passwordless-api.handler';
 export * from './storage/in-memory-store';
 export * from './storage/key-value-store';
 export * from './util/code-challenge-method';
 export * from './util/error-response-factory';
 export * from './util/parsed-json';
 export * from './webid/single-claim-webid-factory';
+export * from './webid/url-claim-webid-factory';
 export * from './webid/webid-factory';
 export * from './webid/webid-profile.handler';
 export * from './webid/webid-response.handler';
+