+++ packages/inrupt-solid-service/CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.15.2...v0.16.0) (2022-03-18)
+
+
+### **Bug Fixes**
+
+* add error handling to getAuthorizationAgents() ([#334](https://github.com/digita-ai/dgt-id-broker/issues/334)) ([7ac2579](https://github.com/digita-ai/dgt-id-broker/commit/7ac25795a4c3c6c2c8c2e181c74ce64e440fe8f7))
+
+
+
+### [0.15.2](https://github.com/digita-ai/dgt-id-broker/compare/v0.15.1...v0.15.2) (2022-02-24)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-service
+
+
+
+
+
+### [0.15.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.15.0...v0.15.1) (2022-02-24)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-service
+
+
+
+
+
+## [0.15.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.14.3...v0.15.0) (2022-02-23)
+
+
+### **Features**
+
+* add getAuthorizationAgents() to SDK ([#333](https://github.com/digita-ai/dgt-id-broker/issues/333)) ([ef5fd85](https://github.com/digita-ai/dgt-id-broker/commit/ef5fd85f720a59505a43b29deb12c8adac6b7024))
+
+
+
+### [0.14.3](https://github.com/digita-ai/dgt-id-broker/compare/v0.14.2...v0.14.3) (2022-02-17)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-service
+
+
+
+
+
+### [0.14.2](https://github.com/digita-ai/dgt-id-broker/compare/v0.14.1...v0.14.2) (2022-02-17)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-service
+
+
+
+
+
+### [0.14.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.14.0...v0.14.1) (2022-02-17)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-service
+
+
+
+
+
+## [0.14.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.13.1...v0.14.0) (2022-02-16)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-service
+
+
+
+
+
 ### [0.13.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.13.0...v0.13.1) (2022-01-14)
 
 **Note:** Version bump only for package @digita-ai/inrupt-solid-service